backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
bcrypt
pydantic
python-dotenv
itsdangerous
python-multipart